# backend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediSync API",
    "description": "Hospital Management System API for managing doctors, patients, and appointments.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/patients/{patientId}/deactivate": {
      "put": {
        "tags": [
          "Patients"
        ],
        "summary": "Deactivate patient",
        "description": "Soft-deletes a patient account. Restricted to Admins.",
        "operationId": "deactivatePatient",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Patient successfully deactivated (No Content)"
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      }
    },
    "/api/patients/{patientId}/activate": {
      "put": {
        "tags": [
          "Patients"
        ],
        "summary": "Activate patient",
        "description": "Restores a deactivated patient account. Restricted to Admins.",
        "operationId": "activatePatient",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Patient successfully activated (No Content)"
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      }
    },
    "/api/patients/{patientId}/": {
      "put": {
        "tags": [
          "Patients"
        ],
        "summary": "Update profile",
        "description": "Allows a patient to update their personal details (address, phone, etc.). Restricted to the account owner.",
        "operationId": "updatePatient",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - You can only update your own profile"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      }
    },
...
    "/api/doctors/{doctorId}/activate": {
      "put": {
        "tags": [
          "Doctors"
        ],
        "summary": "Activate doctor",
        "description": "Restores a deactivated doctor account. Requires ADMIN role.",
        "operationId": "activate",
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Doctor successfully activated (No Content)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role"
          },
          "404": {
            "description": "Doctor not found"
          }
        }
      }
    },
...
    "/api/patients": {
      "post": {
        "tags": [
          "Patients"
        ],
        "summary": "Register new patient",
        "description": "Public endpoint allows new users to create a patient account. No authentication required.",
        "operationId": "registerPatient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRegistrationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (e.g., Email already in use, weak password)"
          }
        },
        "security": []
      }
    },
...
    "/api/patients/{patientId}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Get patient profile",
        "description": "Retrieves public profile information of a patient.",
        "operationId": "getPatientById",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved profile",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found"
          }
        }
      }
    },
    "/api/patients/{patientId}/appointments": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Get patient's appointments",
        "description": "Retrieves the appointment history for a specific patient. Access is restricted to the Patient (owner) or an Admin.",
        "operationId": "getAppointmentsByPatientId",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved appointment history",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Not authorized to view this patient's history"
          },
          "404": {
            "description": "Patient not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PatientUpdateDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "phoneNumber": {
            "type": "string",
            "minLength": 1
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ]
          },
          "allergyIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "county": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        },
        "required": [
          "dateOfBirth",
          "firstName",
          "gender",
          "lastName",
          "phoneNumber"
        ]
      },
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ]
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "county": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MedicalRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "diagnosis": {
            "type": "string"
          },
          "treatmentPlan": {
            "type": "string"
          },
          "prescription": {
            "type": "string"
          },
          "appointmentId": {
            "type": "integer",
            "format": "int64"
          },
          "appointmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "doctorId": {
            "type": "integer",
            "format": "int64"
          },
          "doctorName": {
            "type": "string"
          },
          "patientId": {
            "type": "integer",
            "format": "int64"
          },
          "patientName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DoctorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "specialization": {
            "type": "string"
          },
          "specializationLabel": {
            "type": "string"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentName": {
            "type": "string"
          },
          "appointmentDuration": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AppointmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "appointmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW"
            ]
          },
          "doctor": {
            "$ref": "#/components/schemas/DoctorDto"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "medicalRecord": {
            "$ref": "#/components/schemas/MedicalRecordDto"
          }
        }
      },
...
      "PatientRegistrationDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "pattern": "^(?=.*[0-9])(?=.*[A-Z])(?=.*[!@#$%^&()+=_-])(?=\\S+$).{8,}$"
          },
          "phoneNumber": {
            "type": "string",
            "minLength": 1
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string"
          },
          "allergyIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "county": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        },
        "required": [
          "dateOfBirth",
          "email",
          "firstName",
          "gender",
          "lastName",
          "password",
          "phoneNumber"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}